financeiro/python/requirements.txt
sshtunnel==0.1.5
SQLAlchemy==1.3.5
requests==2.22.0
cx_Oracle==8.0.1
python-decouple==3.3
